# CheckCheck/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CheckCheck REST API",
    "version": "0.1.dev47+ga91abbb.d20260618"
  },
  "paths": {
...
    "/api/checklist/{checklist_id}": {
      "get": {
        "tags": [
          "Checklist"
        ],
        "summary": "Get Checklist",
        "description": "Update existing CheckList",
        "operationId": "get_checklist_api_checklist__checklist_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "checklist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checklist Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckListApiWithSubObj"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Checklist"
        ],
        "summary": "Update Checklist",
        "description": "Update existing CheckList",
        "operationId": "update_checklist_api_checklist__checklist_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "checklist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checklist Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckListUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckListApiWithSubObj"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Checklist"
        ],
        "summary": "Delete Checklist",
        "description": "Delete existing CheckList. The respective users specific CheckListPosition will be deleted as well.",
        "operationId": "delete_checklist_api_checklist__checklist_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "checklist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checklist Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/checklist/{checklist_id}/shares": {
      "get": {
        "tags": [
          "Checklist Sharing"
        ],
        "summary": "List Shares",
        "description": "List all collaborators of a checklist and their permission levels. Owner only.",
        "operationId": "list_shares_api_checklist__checklist_id__shares_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "checklist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checklist Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShareRead"
                  },
                  "title": "Response List Shares Api Checklist  Checklist Id  Shares Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/checklist/{checklist_id}/shares/{user_id}": {
      "put": {
        "tags": [
          "Checklist Sharing"
        ],
        "summary": "Upsert Share",
        "description": "Share the checklist with a user, or update an existing collaborator's permission. Owner only.",
        "operationId": "upsert_share_api_checklist__checklist_id__shares__user_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "checklist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checklist Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Checklist Sharing"
        ],
        "summary": "Delete Share",
        "description": "Revoke a user's access. The owner may revoke anyone; a collaborator may revoke themselves ('leave list').",
        "operationId": "delete_share_api_checklist__checklist_id__shares__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "checklist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checklist Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/checklist/{checklist_id}/transfer-ownership": {
      "post": {
        "tags": [
          "Checklist Sharing"
        ],
        "summary": "Transfer Ownership",
        "description": "Transfer ownership to another user. The previous owner is demoted to an 'edit' collaborator and keeps access. Owner only.",
        "operationId": "transfer_ownership_api_checklist__checklist_id__transfer_ownership_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "checklist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checklist Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferOwnershipRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferOwnershipResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckListApiWithSubObj": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "The display name of the list"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "description": "A text that will be shown at the header"
          },
          "color_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color Id"
          },
          "checked_items_seperated": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Items Seperated",
            "default": true
          },
          "checked_items_collapsed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Items Collapsed",
            "description": "If checked_items_seperated is enabled and this is set to true the client should hide any checked items. ",
            "default": true
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "color": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChecklistColorScheme"
              },
              {
                "type": "null"
              }
            ]
          },
          "position": {
            "$ref": "#/components/schemas/CheckListPositionPublicWithoutChecklistID"
          },
          "labels": {
            "items": {
              "$ref": "#/components/schemas/LabelReadAPI"
            },
            "type": "array",
            "title": "Labels"
          }
        },
        "type": "object",
        "required": [
          "id",
          "color",
          "position",
          "labels"
        ],
        "title": "CheckListApiWithSubObj"
      },
...
      "CheckListPositionPublicWithoutChecklistID": {
        "properties": {
          "index": {
            "type": "number",
            "title": "Index",
            "description": "The position of a checklist relative to other checklist 'index' value. The lowest value should be the first checklist in a client representation."
          },
          "pinned": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned",
            "description": "Pinned items should get a more prominent placement in the client.",
            "default": false
          },
          "archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived",
            "description": "Archived items are filtered out by default. They are considered obsolete.",
            "default": false
          },
          "checked_items_collapsed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Items Collapsed",
            "description": "If true the client should hide checked items in the list view.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "index"
        ],
        "title": "CheckListPositionPublicWithoutChecklistID"
      },
...
      "CheckListUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "The display name of the list"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "description": "A text that will be shown at the header"
          },
          "color_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color Id"
          },
          "checked_items_seperated": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Items Seperated",
            "default": true
          },
          "checked_items_collapsed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Items Collapsed",
            "description": "If checked_items_seperated is enabled and this is set to true the client should hide any checked items. ",
            "default": true
          }
        },
        "type": "object",
        "title": "CheckListUpdate"
      },
...
      "ChecklistColorScheme": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name",
            "description": "Name of the color."
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order",
            "description": "Reserverd for later use. Color order not implemented yet"
          },
          "backgroundcolor_dark_hex": {
            "type": "string",
            "title": "Backgroundcolor Dark Hex"
          },
          "backgroundcolor_light_hex": {
            "type": "string",
            "title": "Backgroundcolor Light Hex"
          },
          "textcolor_dark_hex": {
            "type": "string",
            "title": "Textcolor Dark Hex"
          },
          "textcolor_light_hex": {
            "type": "string",
            "title": "Textcolor Light Hex"
          },
          "accentcolor_dark_hex": {
            "type": "string",
            "title": "Accentcolor Dark Hex"
          },
          "accentcolor_light_hex": {
            "type": "string",
            "title": "Accentcolor Light Hex"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_name",
          "backgroundcolor_dark_hex",
          "backgroundcolor_light_hex",
          "textcolor_dark_hex",
          "textcolor_light_hex",
          "accentcolor_dark_hex",
          "accentcolor_light_hex"
        ],
        "title": "ChecklistColorScheme"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LabelReadAPI": {
        "properties": {
          "color_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "description": "Name of the label."
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order",
            "description": "Label order per user"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          },
          "color": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChecklistColorScheme"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "color"
        ],
        "title": "LabelReadAPI"
      },
...
      "SharePermission": {
        "type": "string",
        "enum": [
          "view",
          "check",
          "edit"
        ],
        "title": "SharePermission",
        "description": "A permission level that can be *granted* to a share target (collaborator or\npublic link). Ownership is not grantable via a share \u2014 it is transferred via\nthe dedicated transfer-ownership action \u2014 so it is intentionally absent here.\nSee ``checkcheckserver.api.access.ChecklistAccessLevel`` for the full ladder\n(which adds ``owner``) used for authorization checks."
      },
      "ShareRead": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "user_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Name"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "permission": {
            "$ref": "#/components/schemas/SharePermission"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "permission"
        ],
        "title": "ShareRead",
        "description": "A single collaborator entry with enough user info to render a share list."
      },
      "ShareUpsertRequest": {
        "properties": {
          "permission": {
            "$ref": "#/components/schemas/SharePermission",
            "description": "Permission level to grant the target user."
          }
        },
        "type": "object",
        "required": [
          "permission"
        ],
        "title": "ShareUpsertRequest"
      },
...
      "TransferOwnershipRequest": {
        "properties": {
          "new_owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "New Owner Id",
            "description": "The user that should become the new owner. The previous owner is demoted to an 'edit' collaborator."
          }
        },
        "type": "object",
        "required": [
          "new_owner_id"
        ],
        "title": "TransferOwnershipRequest"
      },
      "TransferOwnershipResult": {
        "properties": {
          "checklist_id": {
            "type": "string",
            "format": "uuid",
            "title": "Checklist Id"
          },
          "new_owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "New Owner Id"
          },
          "previous_owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Previous Owner Id",
            "description": "The previous owner, now demoted to an 'edit' collaborator."
          }
        },
        "type": "object",
        "required": [
          "checklist_id",
          "new_owner_id",
          "previous_owner_id"
        ],
        "title": "TransferOwnershipResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}